use reduced-size position vector

+++ src/common.rs
@@ -1,5 +1,5 @@
 
 #[derive(PartialEq, Clone, Debug)]
-pub struct Position(pub usize, pub usize);
+pub struct Position(pub u32, pub u32);
 
 pub struct CheckCell(pub Position);
